move process backend into its own library
