+++ smtlib-backends-process/smtlib-backends-process.cabal
@@ -1,0 +1,53 @@
+name:          smtlib-backends-process
+version:       0.1
+synopsis:      An SMT-LIB backend running solvers as external processes.
+description:
+  This library implements an SMT-LIB backend (in the sense of the smtlib-backends package) using by running solvers as external processes.
+
+license:       MIT
+license-file:  LICENSE
+author:        Quentin Aristote
+maintainer:    quentin.aristote@tweag.io
+build-type:    Simple
+category:      SMT
+cabal-version: >=1.10
+
+source-repository head
+  type:     git
+  location: https://github.com/tweag/smtlib-backends
+  subdir:   smtlib-backends-process
+
+source-repository this
+  type:     git
+  location: https://github.com/tweag/smtlib-backends
+  tag:      0.1
+  subdir:   smtlib-backends-process
+
+library
+  hs-source-dirs:   src
+  ghc-options:      -Wall -Wunused-packages
+  exposed-modules:  SMTLIB.Backends.Internal.Process
+  other-extensions: Safe
+  build-depends:
+      async
+    , base           >=4.14 && <4.17.0
+    , bytestring     >=0.10
+    , data-default
+    , typed-process
+
+  default-language: Haskell2010
+
+test-suite test
+  type:             exitcode-stdio-1.0
+  hs-source-dirs:   tests
+  main-is:          Main.hs
+  ghc-options:      -threaded -Wall -Wunused-packages
+  build-depends:
+      base                   >=4.14 && <4.17.0
+    , data-default
+    , smtlib-backends
+    , smtlib-backends-tests
+    , tasty
+    , tasty-hunit
+
+  default-language: Haskell2010
