+++ smtlib-backends.cabal
@@ -32,25 +32,8 @@ library
 
   other-extensions: Safe
   build-depends:
-      async
-    , base           >=4.14 && <4.17.0
-    , bytestring     >=0.10
-    , data-default
-    , typed-process
-
-  default-language: Haskell2010
-
-test-suite test
-  type:             exitcode-stdio-1.0
-  hs-source-dirs:   tests
-  main-is:          Main.hs
-  ghc-options:      -threaded -Wall -Wunused-packages
-  build-depends:
-      base                   >=4.14 && <4.17.0
-    , data-default
-    , smtlib-backends
-    , smtlib-backends-tests
-    , tasty
-    , tasty-hunit
+      base                     >=4.14 && <4.17.0
+    , bytestring               >=0.10
+    , smtlib-backends-process
 
   default-language: Haskell2010
